requirements.txt
Django
django-debug-toolbar
django-allauth
